pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "komatsu-buy-ai"
version = "0.1.0"
description = "Swing buy timing decision support for Komatsu (6301.T)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.0",
  "scikit-learn>=1.4.0",
  "lightgbm>=4.3.0",
  "duckdb>=1.0.0",
  "pyarrow>=15.0.0",
  "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.3.0"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
